requirements.txt
absl-py==2.1.0
dash==2.15.0
dm-sonnet==1.36
dm-tree==0.1.8
flow-vis==0.1
flask==2.2.5
fvcore==0.1.5.post20221221
graph-nets==1.1.0
iopath==0.1.10
imageio==2.31.2
imageio-ffmpeg==0.5.1
matplotlib==3.5.3
numpy==1.21.6
open3d==0.13.0
opencv-python
pandas==1.3.5
plyfile==0.9
protobuf==3.20.3
pyquaternion==0.9.9
pyyaml==6.0.1
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.3
tensorboard==1.15.0
tensorflow-gpu==1.15.0
tensorflow-estimator==1.15.1
tensorflow-probability==0.8.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.67.1
yacs==0.1.8
plotly==5.18.0
openexr
